requirements.txt
flask==3.0.2
openai==1.12.0
azure-identity==1.15.0
azure-monitor-query==1.2.0
pandas==2.2.1
reportlab==4.1.0
python-dotenv==1.0.1
requests==2.31.0